refactor: split send screen input/confirm components
