+++ src/components/send/send-action.tsx
@@ -1,6 +1,9 @@
-import { translate } from "store/index"
-import { InvoiceInput } from "components/pages/send"
+import { useCallback } from "react"
 
+import { translate, useAppDispatcher } from "store/index"
+import useMainQuery from "hooks/use-main-query"
+
+import { ButtonLink } from "components/link"
 import SendIntraLedgerAction, {
   SendIntraLedgerActionProps,
 } from "components/send/send-intra-ledger-action"
@@ -13,51 +16,98 @@ import {
 import SendOnChainAction, {
   SendOnChainActionProps,
 } from "components/send/send-onchain-action"
+import { SendScreenInput } from "components/pages/send"
+import { Spinner } from "@galoymoney/react"
 
-export type SendActionProps = InvoiceInput & {
+export type SendActionProps = SendScreenInput & {
   btcWalletId: string
   btcWalletBalance: number
   reset: () => void
 }
 
-type FCT = React.FC<SendActionProps>
+type FCT = React.FC<{ children?: React.ReactNode; input: SendScreenInput }>
 
-const SendAction: FCT = (props) => {
-  let { errorMessage } = props
+const SendAction: FCT = ({ children, input }) => {
+  const dispatch = useAppDispatcher()
+  const { btcWalletId, btcWalletBalance } = useMainQuery()
 
-  if (props.satAmount && props.satAmount > props.btcWalletBalance) {
-    errorMessage = translate("Payment amount exceeds balance of %{balance} sats", {
-      balance: props.btcWalletBalance,
-    })
+  const reset = useCallback(() => {
+    dispatch({ type: "navigate", path: "/send" })
+  }, [dispatch])
+
+  if (!btcWalletId) {
+    return <ButtonLink to="/login">{translate("Login to send")}</ButtonLink>
   }
 
-  if (errorMessage) {
+  if (input.errorMessage) {
+    return <div className="error">{input.errorMessage}</div>
+  }
+
+  const inputPending =
+    input.amount === undefined ||
+    input.destination === undefined ||
+    input.memo === undefined
+
+  const satAmountPending =
+    typeof input.amount === "number" && input.satAmount === undefined
+
+  if (input.satAmount && input.satAmount > btcWalletBalance) {
+    const errorMessage = translate("Payment amount exceeds balance of %{balance} sats", {
+      balance: btcWalletBalance,
+    })
     return <div className="error">{errorMessage}</div>
   }
 
+  if (children) {
+    return <>{children}</>
+  }
+
+  const showSpinner = inputPending || satAmountPending
+
+  if (showSpinner) {
+    return <Spinner size="big" />
+  }
+
+  if (input.satAmount && input.satAmount > btcWalletBalance) {
+    return (
+      <div className="error">
+        {translate("Payment amount exceeds balance of %{balance} sats", {
+          balance: btcWalletBalance,
+        })}
+      </div>
+    )
+  }
+
   const validInput =
-    props.valid &&
-    (props.fixedAmount || typeof props.amount === "number") &&
-    (props.paymentType !== "intraledger" || props.recipientWalletId)
+    input.valid &&
+    (input.fixedAmount || typeof input.amount === "number") &&
+    (input.paymentType !== "intraledger" || input.recipientWalletId)
 
   if (!validInput) {
     return <button disabled>{translate("Enter amount and destination")}</button>
   }
 
-  if (props.paymentType === "lightning") {
-    return props.fixedAmount ? (
-      <SendLnInvoiceAction {...(props as SendLnActionProps)} />
+  const sendActionProps = {
+    ...input,
+    btcWalletId,
+    btcWalletBalance,
+    reset,
+  }
+
+  if (sendActionProps.paymentType === "lightning") {
+    return sendActionProps.fixedAmount ? (
+      <SendLnInvoiceAction {...(sendActionProps as SendLnActionProps)} />
     ) : (
-      <SendLnNoAmountInvoiceAction {...(props as SendLnNoAmountActionProps)} />
+      <SendLnNoAmountInvoiceAction {...(sendActionProps as SendLnNoAmountActionProps)} />
     )
   }
 
-  if (props.paymentType === "onchain") {
-    return <SendOnChainAction {...(props as SendOnChainActionProps)} />
+  if (sendActionProps.paymentType === "onchain") {
+    return <SendOnChainAction {...(sendActionProps as SendOnChainActionProps)} />
   }
 
-  if (props.paymentType === "intraledger") {
-    return <SendIntraLedgerAction {...(props as SendIntraLedgerActionProps)} />
+  if (sendActionProps.paymentType === "intraledger") {
+    return <SendIntraLedgerAction {...(sendActionProps as SendIntraLedgerActionProps)} />
   }
 
   return <button disabled>{translate("Enter amount and destination")}</button>
